src/requirements.txt
colorama
pyfiglet
PySocks
requests
termcolor
urllib3
youtube-dl
